getting ready for release
