--- data/colours
@@ -1,12 +1,0 @@
-Nematoda,#48a365
-Tardigrada,#ffffff
-Arthropoda,#d0694a
-Proteobacteria,#ffb917
-Actinobacteria,#926eb3
-Ascomycota,#a6cee3
-no-hit,#d3d3d3
-Chordata,#ec84ba
-Cnidaria,#e0d799
-Platyhelminthes,#b89c75
-Bacteriodetes,#fdb761
-other,#ffffff
