--- data/refcov
@@ -1,1 +1,0 @@
-cas0,84660821,75823917
