--- test_data/colours
@@ -1,4 +1,0 @@
-Nematoda,#48a365
-Tardigrada,#48a365
-Actinobacteria,#926eb3
-other,#ffffff
