--- test_data/test1.nHa.duped.filtered.fna
@@ -1,8 +1,0 @@
->contig_1
-TTTTTTATTTTTTGTTTAAAAGACGCATTTAATTGTCTTAGTTTTTTCTTAACGTTTCGCATTCCCAATCAATTGTATCCAGAATTTAAATGTGCTTTTTCACACGGCGATTTTATGGCACATAAATTAAGGGAATCAAAACTCATAATCATCGATGAAGTCAGCATGCTTCATAAAAATGTAATCTTTGCTATCGACTCAACTTTAAGGTTTTTTTTATTTAATTTTTTTAATTAAAAACTTTTAACTATAAAACATAATCAAAAATTTTTTTATATCTTATACACATATTATTATATTCATCAATTTACATTCTAAAATTAGATAAATACTATTTAAATTTTAAAATAAAATTTTTAAAAATAAAATTAAAAATTTTAGTAGCTTTAAATTTTTTTATTTTAGATACTTATATTCTGCAAGCGACATTAAACGAAAAATCCCATTTGCAAGAAAAACAATTATATTTGGAGGAGATTGGAAACAGCTGCTCCCTGTTATTCCAGGAAGCAATATTAATCAACAAGCACTTGCAAGCTTCAAATTTACTAAGTTCTTCAACCTCTGTCAACCAATAACATTGCAAAAAAACATGAAAGTTAATAAAGATGACGGCGATTTTGCACTGTTTCTTATATTAATTGGAAATGGATTAATTGGAACAGAGGATAAAAATGGTGTCTTTGAACTTCCAACTCAAATGATAGCAAATAAGCTTAAAGAATTAGTAGATTTTTGCTTCGACACCGATTGCTT
->contig_1
-TTTTTTATTTTTTGTTTAAAAGACGCATTTAATTGTCTTAGTTTTTTCTTAACGTTTCGCATTCCCAATCAATTGTATCCAGAATTTAAATGTGCTTTTTCACACGGCGATTTTATGGCACATAAATTAAGGGAATCAAAACTCATAATCATCGATGAAGTCAGCATGCTTCATAAAAATGTAATCTTTGCTATCGACTCAACTTTAAGGTTTTTTTTATTTAATTTTTTTAATTAAAAACTTTTAACTATAAAACATAATCAAAAATTTTTTTATATCTTATACACATATTATTATATTCATCAATTTACATTCTAAAATTAGATAAATACTATTTAAATTTTAAAATAAAATTTTTAAAAATAAAATTAAAAATTTTAGTAGCTTTAAATTTTTTTATTTTAGATACTTATATTCTGCAAGCGACATTAAACGAAAAATCCCATTTGCAAGAAAAACAATTATATTTGGAGGAGATTGGAAACAGCTGCTCCCTGTTATTCCAGGAAGCAATATTAATCAACAAGCACTTGCAAGCTTCAAATTTACTAAGTTCTTCAACCTCTGTCAACCAATAACATTGCAAAAAAACATGAAAGTTAATAAAGATGACGGCGATTTTGCACTGTTTCTTATATTAATTGGAAATGGATTAATTGGAACAGAGGATAAAAATGGTGTCTTTGAACTTCCAACTCAAATGATAGCAAATAAGCTTAAAGAATTAGTAGATTTTTGCTTCGACACCGATTGCTT
->contig_2
-TAGCAGAAAGAAAAATAAAAAATTTTAATGAAGATCAAAAAAACGCATTTGATAAAGTAATTTATATTTAATTAAAAAAAAAATATATTAAAATAATTTAGATTATGGAATCAGTAAGAAGTAAAGAAATTGACCATAATGGCAAACTATTTATGTTAACAGGAGATGGAGGCACAGGTAAATTTCATATATTTAATTAATATTTTTAAATTAATTTTTAGGAAAATCCTATTTATATAATGGATTGATAGCTCAGTGCAAAGCTCATCAATTTGAAGTTATTCCATGCGCATCAACTGGAATAGCTTCAATATTATTATTGTGCAATTTTGGTACAGCACACAGAACGTTTCGCATTCCCAATCAATTGTATCCAGAATTTAAATGTGCTTTTTCACACGGCGATTTTATGGCACATAAATTAAGGGAATCAAAACTTATAATCATCGATGAAGTCAGCATGCTTCATAAAAATGTAATCTTTGCTATCGACTCAACTTTAAGGTTTTTTTTATTTAATTTTTTTAATTAAAAACTTTCAACTATAAAACATAATCAAAAATTTTTTTATATCTTATACACATGTTCATTAATTTACATTCTAAAATTAGATAAATACTATTTAAATTTTAAAATAAAATTTTTAAAAATAAAATTAAAAATTTTAGTAGCTTTAAATTTTTTTATTTTAGATACTTATATTCTGCAAGCGACATTAAACGAAAAATCCCATTTGCAGGAAAAACAATTATATTTGGAGGAGATTGGAAACAGCTGCTCCCTGTTGTTCCAGGAAGCAATATTAATCAACAAGCACTTGCAAGCTTCAAATTTACTAAGTTCTTCAACCTCTGTCAACCAATAACATTGCAAAAAAACATGAGAGTTAATAAAGATGAAGGCGTTTTTGCAATGTTTCTTAAATTGAATGGAAATGGATTAATTGGAACAGAGGATAAAAATGGTGTCTTTGAAATACCACCTAAAATGATAGCTAAAAATCTTATAGAATTGGTAGATTTTTGCTTTGACACCGATTGTTTACGCAATCATACCGTAT
->contig_2
-AGATAGATAGAT
