--- test_data/test1.nHa.filtered.fna
@@ -1,4 +1,0 @@
->contig_1
-TTTTTTATTTTTTGTTTAAAAGACGCATTTAATTGTCTTAGTTTTTTCTTAACGTTTCGCATTCCCAATCAATTGTATCCAGAATTTAAATGTGCTTTTTCACACGGCGATTTTATGGCACATAAATTAAGGGAATCAAAACTCATAATCATCGATGAAGTCAGCATGCTTCATAAAAATGTAATCTTTGCTATCGACTCAACTTTAAGGTTTTTTTTATTTAATTTTTTTAATTAAAAACTTTTAACTATAAAACATAATCAAAAATTTTTTTATATCTTATACACATATTATTATATTCATCAATTTACATTCTAAAATTAGATAAATACTATTTAAATTTTAAAATAAAATTTTTAAAAATAAAATTAAAAATTTTAGTAGCTTTAAATTTTTTTATTTTAGATACTTATATTCTGCAAGCGACATTAAACGAAAAATCCCATTTGCAAGAAAAACAATTATATTTGGAGGAGATTGGAAACAGCTGCTCCCTGTTATTCCAGGAAGCAATATTAATCAACAAGCACTTGCAAGCTTCAAATTTACTAAGTTCTTCAACCTCTGTCAACCAATAACATTGCAAAAAAACATGAAAGTTAATAAAGATGACGGCGATTTTGCACTGTTTCTTATATTAATTGGAAATGGATTAATTGGAACAGAGGATAAAAATGGTGTCTTTGAACTTCCAACTCAAATGATAGCAAATAAGCTTAAAGAATTAGTAGATTTTTGCTTCGACACCGATTGCTT
->contig_2
-TAGCAGAAAGAAAAATAAAAAATTTTAATGAAGATCAAAAAAACGCATTTGATAAAGTAATTTATATTTAATTAAAAAAAAAATATATTAAAATAATTTAGATTATGGAATCAGTAAGAAGTAAAGAAATTGACCATAATGGCAAACTATTTATGTTAACAGGAGATGGAGGCACAGGTAAATTTCATATATTTAATTAATATTTTTAAATTAATTTTTAGGAAAATCCTATTTATATAATGGATTGATAGCTCAGTGCAAAGCTCATCAATTTGAAGTTATTCCATGCGCATCAACTGGAATAGCTTCAATATTATTATTGTGCAATTTTGGTACAGCACACAGAACGTTTCGCATTCCCAATCAATTGTATCCAGAATTTAAATGTGCTTTTTCACACGGCGATTTTATGGCACATAAATTAAGGGAATCAAAACTTATAATCATCGATGAAGTCAGCATGCTTCATAAAAATGTAATCTTTGCTATCGACTCAACTTTAAGGTTTTTTTTATTTAATTTTTTTAATTAAAAACTTTCAACTATAAAACATAATCAAAAATTTTTTTATATCTTATACACATGTTCATTAATTTACATTCTAAAATTAGATAAATACTATTTAAATTTTAAAATAAAATTTTTAAAAATAAAATTAAAAATTTTAGTAGCTTTAAATTTTTTTATTTTAGATACTTATATTCTGCAAGCGACATTAAACGAAAAATCCCATTTGCAGGAAAAACAATTATATTTGGAGGAGATTGGAAACAGCTGCTCCCTGTTGTTCCAGGAAGCAATATTAATCAACAAGCACTTGCAAGCTTCAAATTTACTAAGTTCTTCAACCTCTGTCAACCAATAACATTGCAAAAAAACATGAGAGTTAATAAAGATGAAGGCGTTTTTGCAATGTTTCTTAAATTGAATGGAAATGGATTAATTGGAACAGAGGATAAAAATGGTGTCTTTGAAATACCACCTAAAATGATAGCTAAAAATCTTATAGAATTGGTAGATTTTTGCTTTGACACCGATTGTTTACGCAATCATACCGTAT
