+++ test_files/test.assembly.list.txt
@@ -1,0 +1,2 @@
+contig_1
+contig_2
