+++ test_files/test.diamond.out
@@ -1,0 +1,14 @@
+contig_1    232323  200
+contig_2    232323  500
+contig_2    232323  1000
+contig_2    232323  500
+contig_2    979556  300
+contig_3    979556  10000
+contig_4    979556  1000
+contig_5    6252    1000
+contig_6    232323  1000
+contig_6    6252    1000
+contig_6    979556  1000
+contig_6    232323  1000
+contig_7    6252    1000
+contig_9    6252    100
