+++ test_files/test.mapping_2.bam.cov
@@ -1,0 +1,16 @@
+## blobtools v0.9.19
+## Total Reads = 37278
+## Mapped Reads = 37278
+## Unmapped Reads = 0
+## Source(s) : /Users/dominik/git/blobtools/test_data/test.mapping_2.bam
+# contig_id	read_cov	base_cov
+contig_1	489	141.55026455
+contig_2	622	137.903773585
+contig_3	182	60.3488372093
+contig_4	6001	1338.51840168
+contig_5	1129	408.280130293
+contig_6	80	50.4675925926
+contig_7	739	39.1736453202
+contig_8	1497	154.437766411
+contig_9	725	105.721701063
+contig_10	25814	965.662681333
