+++ test_files/test.refcov.txt
@@ -1,0 +1,5 @@
+cov0,15313,15300
+bam0,15313,15300
+cov1,37278,15300
+bam1,37278,15300
+covsum,52591,30600
